Revert "paging scheme"
This reverts commit d12de7d4
