+++ shenyu-admin/src/http/http-debug-app-auth-controller-api.http
@@ -17,54 +17,11 @@
 
 # if you debug api, replace your own token
 
-### app auth list search
-POST http://localhost:9095/appAuth/list/search
-Accept: application/json
-Content-Type: application/json
-X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyTmFtZSI6ImFkbWluIiwiZXhwIjoxNjQ4NjUwMDg2fQ.aDeChT_Ey6FwYDdzSkc9ZLBHd5v-LVUZ6BPcYqJCo-Y
-
-{
-  "pageNum": 1,
-  "pageSize": 12,
-  "condition": {
-    "appKey": "9ABE7111A59B4F9CA9A5BA5BB5973F6F"
-  }
-}
-
-### app auth list search adaptor
-POST http://localhost:9095/appAuth/list/search/adaptor
-Accept: application/json
-Content-Type: application/json
-X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyTmFtZSI6ImFkbWluIiwiZXhwIjoxNjQ4NjUwMDg2fQ.aDeChT_Ey6FwYDdzSkc9ZLBHd5v-LVUZ6BPcYqJCo-Y
-
-{
-  "pageNum": 1,
-  "pageSize": 12,
-  "condition": {
-  }
-}
-
-### updateSk
-POST http://localhost:9095/appAuth/apply
-Accept: application/json
-Content-Type: application/json
-X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyTmFtZSI6ImFkbWluIiwiZXhwIjoxNjQ4NjUwMDg2fQ.aDeChT_Ey6FwYDdzSkc9ZLBHd5v-LVUZ6BPcYqJCo-Y
-
-{
-  "appName": "appName",
-  "userId": 123,
-  "open": true,
-  "phone": "189827364",
-  "pathList": [
-    "test"
-  ]
-}
-
 ### updateSk
 GET http://localhost:9095/appAuth/updateSk?appKey=123&appSecret=123
 Accept: application/json
 Content-Type: application/json
-X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyTmFtZSI6ImFkbWluIiwiZXhwIjoxNjQ4NjUwMDg2fQ.aDeChT_Ey6FwYDdzSkc9ZLBHd5v-LVUZ6BPcYqJCo-Y
+X-Access-Token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyTmFtZSI6ImFkbWluIiwiZXhwIjoxNjQ2MzIyMTI4fQ.amilvuSJfMM-qt9bYsEr-Ih12NKC9aK-3mB-VUq-fdo
 
 {
   "id": 123,
