+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/model/result/ShenyuAdminResult.java
@@ -21,11 +21,20 @@ import org.apache.shenyu.common.exception.CommonErrorCode;
 import org.springframework.http.HttpStatus;
 
 import java.io.Serializable;
+import java.util.Objects;
 
 /**
  * ShenyuAdminResult.
  */
-public class ShenyuAdminResult extends AdminResult<Object> implements Serializable {
+public class ShenyuAdminResult implements Serializable {
+
+    private static final long serialVersionUID = -2792556188993845048L;
+
+    private Integer code;
+
+    private String message;
+
+    private Object data;
 
     /**
      * Instantiates a new shenyu result.
@@ -42,7 +51,9 @@ public class ShenyuAdminResult extends AdminResult<Object> implements Serializab
      * @param data    the data
      */
     public ShenyuAdminResult(final Integer code, final String message, final Object data) {
-        super(code, message, data);
+        this.code = code;
+        this.message = message;
+        this.data = data;
     }
 
     /**
@@ -119,5 +130,84 @@ public class ShenyuAdminResult extends AdminResult<Object> implements Serializab
     private static ShenyuAdminResult get(final int code, final String msg, final Object data) {
         return new ShenyuAdminResult(code, msg, data);
     }
-    
+
+    /**
+     * Gets the value of code.
+     *
+     * @return the value of code
+     */
+    public Integer getCode() {
+        return code;
+    }
+
+    /**
+     * Sets the code.
+     *
+     * @param code code
+     */
+    public void setCode(final Integer code) {
+        this.code = code;
+    }
+
+    /**
+     * Gets the value of message.
+     *
+     * @return the value of message
+     */
+    public String getMessage() {
+        return message;
+    }
+
+    /**
+     * Sets the message.
+     *
+     * @param message message
+     */
+    public void setMessage(final String message) {
+        this.message = message;
+    }
+
+    /**
+     * Gets the value of data.
+     *
+     * @return the value of data
+     */
+    public Object getData() {
+        return data;
+    }
+
+    /**
+     * Sets the data.
+     *
+     * @param data data
+     */
+    public void setData(final Object data) {
+        this.data = data;
+    }
+
+    @Override
+    public String toString() {
+        return "ShenyuAdminResult{"
+                + "code=" + code
+                + ", message='" + message
+                + '\'' + ", data=" + data
+                + '}';
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof ShenyuAdminResult)) {
+            return false;
+        }
+        ShenyuAdminResult that = (ShenyuAdminResult) o;
+        return Objects.equals(code, that.code) && Objects.equals(message, that.message) && Objects.equals(data, that.data);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(code, message, data);
+    }
 }
