+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/service/AppAuthService.java
@@ -32,7 +32,7 @@ import java.util.List;
 /**
  * this is application authority service.
  */
-public interface AppAuthService extends PageService<AppAuthQuery, AppAuthVO> {
+public interface AppAuthService {
 
 
     /**
